LayoutTests/http/tests/site-isolation/history/add-iframe-while-changing-document-title.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifies that adding an iframe while changing the document title does not increase history.length");
jsTestIsAsync = true;

let title = true;
function changeTitle() {
    document.title = title ? "a" : "b";
    title = !title;
}

onmessage = (event) => {
    shouldBe("event.data", "'onload'");
    shouldBe("history.length", "1");
    finishJSTest();
}

onload = async () => {
    await testRunner?.clearBackForwardList();

    setInterval(changeTitle, 10);

    const iframe = document.createElement('iframe');
    iframe.src = 'http://localhost:8000/site-isolation/resources/post-message-from-child-to-parent.html';
    document.body.appendChild(iframe);
}
</script>
